업데이트 BackEnd/src/main/java/com/sai/model/service/FeedServiceImpl.java

+++ BackEnd/src/main/java/com/sai/model/service/FeedServiceImpl.java
@@ -304,16 +304,27 @@ public class FeedServiceImpl implements FeedService {
 			}
 		}
 
-		// 투표 삭제 후 재생성
+			// 투표 삭제 후 재생성
 		if (updateBoardRequestDto.isPollModified()) {
 			// 삭제
 			Poll poll = pollRepository.findByBoardBoardId(board.getBoardId());
-			pollRepository.delete(poll);
 
-			// 재생성
+			// 투표가 있는 경우
+			if (poll != null) {
+				pollRepository.delete(poll);
+
+				// 재생성
+				PollRequest pollRequest = updateBoardRequestDto.getPollRequest();
+				if (pollRequest != null)
+					pollService.createPoll(pollRequest);
+			}
+		}
+		// 투표가 없는 경우 새로 만들기
+		else {
 			PollRequest pollRequest = updateBoardRequestDto.getPollRequest();
-			if (pollRequest != null)
+			if (pollRequest != null) {
 				pollService.createPoll(pollRequest);
+			}
 		}
 
 		// 태그 삭제 후 재생성
